pyproject.toml
[tool.poetry]
name = "tweetprint"
version = "0.1.0"
description = ""
authors = ["tasbasmn <nazifts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.22.0"
reportlab = "^4.2.2"
pillow = "^10.4.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
